don't fail if files don't exist

+++ scripts/dm_compile_run.sh
@@ -45,4 +45,4 @@ else
 	fi
 fi
 
-rm "$file".*
+[[ -e "$file.dme" ]] && rm "$file".*
